% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_models.R
\name{ensemble_models}
\alias{ensemble_models}
\title{Retrieve ensemble weather forecasts from the Open-Meteo API}
\usage{
ensemble_models(
  location,
  start = NULL,
  end = NULL,
  hourly = NULL,
  daily = NULL,
  response_units = NULL,
  model = NULL,
  timezone = "auto"
)
}
\arguments{
\item{location}{Required. The location for which data will be retrieved.
Supplied as either a \code{c(latitude,longitude)} WGS84 coordinate pair or a
place name string (with co-ordinates obtained via \code{\link[=geocode]{geocode()}}).}

\item{start, end}{Start and end dates in ISO 8601 (e.g. "2020-12-31"). If no
dates are supplied, data for the next 7 days will be provided by default.}

\item{hourly, daily}{At least one required. A marine weather variable accepted
by the API, or list thereof. See details below.}

\item{response_units}{Supply to convert response units for wave heights. This
defaults to: \verb{list(length_unit="metric") for meters. Specify "Imperial" for feet.}}

\item{model}{Supply to specify an ensemble model for forecasted values (refer
to the API documentation).}

\item{timezone}{specify timezone for time data as a string, i.e.
"australia/sydney" (defaults to "auto", the timezone local to the specified
\code{location}).}
}
\value{
Requested ensemble model data for the given location and time, as
a tidy tibble.
}
\description{
\code{ensemble_models()} calls the Open-Meteo Ensemble Models API to obtain
meteorological ensemble member forecasts for a given location. Limited
historical data is also available via this API.

Refer to the API documentation at: \url{https://open-meteo.com/en/docs/ensemble-api}
}
\details{
You will need to specify at least one variable to retrieve, such as
temperature, that you want data for. These variables are sampled or
aggregated at \emph{hourly} intervals, and can be supplied as a list to request
multiple variables over the same time period.

Example \emph{Hourly} forecast variables include:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{temperature_2m} \tab Air temperature at 2 meters above ground \cr
   \code{precipitation} \tab Sum of rain, showers, and snow over the preceding hour \cr
   \code{windspeed_10m} \tab Wind speed at 10 meters above ground \cr
   \code{cloudcover} \tab Total cloud cover as an area fraction \cr
   \code{pressure_msl} \tab Atmospheric air pressure at mean sea level \cr
}


Different ensemble models can be specified, which may differ in the
weather variables forecasted. Models include:\tabular{lll}{
   \strong{Model} \tab \strong{Origin} \tab \strong{Resolution} \cr
   \code{icon_global} \tab Germany \tab 26 km \cr
   \code{gfs05} \tab USA \tab 50 km \cr
   \code{gem_global} \tab Canada \tab 25 km \cr
}


For all models and their available fields, refer to the full documentation
for the ensemble models API at: \url{https://open-meteo.com/en/docs/ensemble-api}
}
\examples{
\donttest{
# Obtain temperature ensemble forecasts for Indianapolis in Imperial units
# over the next 7 days in Fahrenheit, with GFS
ensemble_models("Indianapolis",
  hourly = "apparent_temperature",
  response_units = list(temperature_unit = "fahrenheit"),
  model = "gfs_seamless"
)
}
}
