% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Retrieve Data from the Open-Meteo Geocoding API}
\usage{
geocode(location_string, n_results = 1, language = "en")
}
\arguments{
\item{location_string}{location name to search for via fuzzy matching}

\item{n_results}{number of locations provided in response (up to 100)}

\item{language}{response language (lower-cased two-letter string)}
}
\value{
details for each location (latitude, longitude, elevation ...)
}
\description{
\code{geocode()} calls the Open-Meteo Geocoding API. API documentation is
available at \url{https://open-meteo.com/en/docs/geocoding-api}
}
\examples{
dt <- geocode("Sydney")
sydney_coords <- c(dt$latitude, dt$longitude)
}
