% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_forecast.R
\name{weather_forecast}
\alias{weather_forecast}
\title{Retrieve Weather Forecasts from the Open-Meteo Forecasting API}
\usage{
weather_forecast(
  location,
  start = NULL,
  end = NULL,
  hourly = NULL,
  daily = NULL,
  response_units = NULL,
  model = NULL,
  timezone = "auto"
)
}
\arguments{
\item{location}{\code{c(latitude,longitude)} WGS84 coordinate pair or a place name (via a fuzzy search using \code{\link[=geocode]{geocode()}})}

\item{start}{start date in ISO 8601 (e.g. "2020-12-30")}

\item{end}{end date in ISO 8601 (e.g. "2021-01-30")}

\item{hourly}{an hourly weather variable provided by the API, or list thereof}

\item{daily}{a daily weather variable provided by the API, or list thereof}

\item{response_units}{convert temperature, windspeed, or precipitation units, defaulting to:
\code{c(temperature_unit = "celsius", windspeed_unit = "kmh", precipitation_unit = "mm")}}

\item{model}{specify a model (or list of models) for forecasted values
(defaults to autoselection of best model)}

\item{timezone}{specify timezone of data (defaults to the timezone local to the specified \code{location})}
}
\value{
specified weather forecast data for the given location and time
}
\description{
\code{weather_forecast()} calls the Open-Meteo Weather Forecast API for a given
location. Location provided either as string or \code{c(latitude,longitude)}. The
next 7 days will be provided if start and end days are not specified.
Full API documentation is available at: \url{https://open-meteo.com/en/docs}.

Call weather_variables()] to retrieve possible values -- unsure if possible atm, maybe from \url{https://github.com/open-meteo/open-meteo/blob/main/openapi.yml}
}
\examples{

# obtain temperature forecasts for London's next 7 days
weather_forecast("London",hourly="temperature_2m")

}
