% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_now.R
\name{weather_now}
\alias{weather_now}
\title{Retrieve Current Weather from the Open-Meteo API}
\usage{
weather_now(location, response_units = NULL, timezone = "auto")
}
\arguments{
\item{location}{\code{c(latitude,longitude)} WGS84 coordinate pair or a place name (via a fuzzy search using \code{\link[=geocode]{geocode()}})}

\item{response_units}{convert temperature, windspeed, or precipitation units, defaulting to:
\code{c(temperature_unit = "celsius", windspeed_unit = "kmh", precipitation_unit = "mm")}}

\item{timezone}{specify timezone of data (defaults to the timezone local to the specified \code{location})}
}
\value{
Current weather conditions: temperature, windspeed, winddirection and weathercode
}
\description{
\code{weather_now()} calls the Open-Meteo Weather API to return current weather
for a given location. Location provided either as string or \code{c(latitude,longitude)}.

Full API documentation is available at: \url{https://open-meteo.com/en/docs}.
}
\examples{

# obtain temperature forecasts for London's next 7 days
weather_forecast("London",hourly="temperature_2m")

}
