% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_variables.R
\name{weather_variables}
\alias{weather_variables}
\title{Retrieve valid hourly and daily weather variables}
\usage{
weather_variables()
}
\value{
A list of valid hourly and daily weather variables
}
\description{
\code{weather_variables()} retrieves an incomplete list of \emph{hourly} and \emph{daily}
variables accepted by \code{\link[=weather_forecast]{weather_forecast()}} and \code{\link[=weather_history]{weather_history()}}, such as
temperature or precipitation.

Refer to the following documentation for the forecast and history API
endpoints for detailed descriptions, units, and caveats:

Forecast API \url{https://open-meteo.com/en/docs}

Historical API \url{https://open-meteo.com/en/docs/historical-weather-api}
}
\examples{
\dontrun{
weather_variables()
}
}
