% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_variables.R
\name{weather_variables}
\alias{weather_variables}
\title{Retrieve a List of Weather Variables Accepted by the Open-Meteo Forecasting API}
\usage{
weather_variables()
}
\value{
a list of weather variables
}
\description{
\code{weather_variables()} retrieves a list of  hourly and daily variables accepted
by \code{\link[=weather_forecast]{weather_forecast()}} and \code{\link[=weather_history]{weather_history()}}, such as temperature and precipitation.

Refer to the following documentation for the forecast and history API
endpoints for detailed descriptions, units, and caveats.

Forecast: \url{https://open-meteo.com/en/docs}
Historical: \url{https://open-meteo.com/en/docs/historical-weather-api}
}
\examples{

weather_variables()

}
